Reorganise file names
