+++ lib/ll.rb
@@ -1,0 +1,5 @@
+require 'll/wk'
+require 'll/wk/api/version'
+# Base LL Module
+module LL
+end
