+++ lib/ll/wk.rb
@@ -1,0 +1,6 @@
+require 'll/wk/api'
+module LL
+  # Module for WebKiosk functions
+  module WK
+  end
+end
