+++ lib/ll/wk/api/version.rb
@@ -1,0 +1,7 @@
+module LL
+  module WK
+    module API
+      VERSION = "0.1.0"
+    end
+  end
+end
