--- lib/llwk_api/version.rb
@@ -1,3 +1,0 @@
-module LlwkApi
-  VERSION = "0.1.0"
-end
