+++ spec/llwk_api_spec.rb
@@ -1,9 +1,5 @@
-RSpec.describe LlwkApi do
-  it "has a version number" do
-    expect(LlwkApi::VERSION).not_to be nil
-  end
-
-  it "does something useful" do
-    expect(false).to eq(true)
+RSpec.describe LL::WK::API do
+  it 'has a version number' do
+    expect(LL::WK::API::VERSION).not_to be nil
   end
 end
