+++ spec/spec_helper.rb
@@ -1,9 +1,9 @@
-require "bundler/setup"
-require "llwk_api"
+require 'bundler/setup'
+require 'll'
 
 RSpec.configure do |config|
   # Enable flags like --only-failures and --next-failure
-  config.example_status_persistence_file_path = ".rspec_status"
+  config.example_status_persistence_file_path = '.rspec_status'
 
   # Disable RSpec exposing methods globally on `Module` and `main`
   config.disable_monkey_patching!
